- change int to float in GameSetting

+++ Game/Assets/PumpDevFolder/Script/GameSetting.cs
@@ -5,7 +5,7 @@ public class GameSetting : MonoBehaviour {
 
 	public static int Player1Health;
 	public static int Player2Health;
-	public static int Player1Speed;
-	public static int Player2Speed;
+	public static float Player1Speed;
+	public static float Player2Speed;
 
 }
